milan: merge with avb stream.c/.h
There are some #ifdef's now, but it seems to be possible to
move the aaf vs iec61883 specifics into their own files (that already
exist).
